refactor: TimeZoneConfig 수정
