--- src/main/java/com/lucy/arti/designer/dto/DesignerRequestDto.java
@@ -1,4 +1,0 @@
-package com.lucy.arti.designer.dto;
-
-public class DesignerRequestDto {
-}
