+++ src/main/java/com/lucy/arti/global/config/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.lucy.arti.config;
+package com.lucy.arti.global.config;
 
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
