+++ src/main/java/com/lucy/arti/member/controller/MemberController.java
@@ -42,6 +42,4 @@ public class MemberController {
     public ResponseEntity<?> saveList(final Authentication authentication) {
         return ResponseEntity.ok(memberService.saveListShow(authentication));
     }
-
-
 }
