--- src/main/java/com/lucy/arti/oauth/dto/AuthCode.java
@@ -1,13 +1,0 @@
-package com.lucy.arti.oauth.dto;
-
-public class AuthCode {
-    private String authCode;
-
-    public String getAuthCode() {
-        return authCode;
-    }
-
-    public void setAuthCode(String authCode) {
-        this.authCode = authCode;
-    }
-}
