+++ src/main/java/com/lucy/arti/oauth/dto/Authority.java
@@ -19,9 +19,5 @@ public class Authority {
     @Column(name = "authority_name", length = 50)
     @Enumerated(EnumType.STRING)
     private UserRole authorityName;
-
-    public String getAuthorityName(){
-        return this.authorityName.toString();
-    }
 }
