+++ src/main/java/com/lucy/arti/oauth/dto/KakaoOAuth2Properties.java
@@ -1,20 +1,10 @@
 package com.lucy.arti.oauth.dto;
 
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
 @Component
-//@ConfigurationProperties(prefix = "spring.security.oauth2.client.registration.kakao")
 public class KakaoOAuth2Properties {
     @Value("${spring.security.oauth2.client.registration.kakao.client-secret}")
     private String clientSecret;
-
-    public String getClientSecret() {
-        return clientSecret;
-    }
-
-    public void setClientSecret(String clientSecret) {
-        this.clientSecret = clientSecret;
-    }
 }
