+++ src/main/java/com/lucy/arti/view/domain/View.java
@@ -2,7 +2,6 @@ package com.lucy.arti.view.domain;
 
 import com.lucy.arti.clothes.domain.Clothes;
 import com.lucy.arti.member.domain.Member;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
