+++ src/main/java/com/lucy/arti/winner/repository/WinnerRepository.java
@@ -3,6 +3,6 @@ package com.lucy.arti.winner.repository;
 import com.lucy.arti.winner.domain.Winner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WinnerRepository extends JpaRepository<Winner, Long>, WinnerRepositoryCustom {
+public interface WinnerRepository extends JpaRepository<Winner, Long> {
 
 }
