--- src/main/java/com/lucy/arti/winner/repository/WinnerRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.lucy.arti.winner.repository;
-
-public interface WinnerRepositoryCustom {
-}
