--- src/main/java/com/lucy/arti/winner/repository/WinnerRepositoryImpl.java
@@ -1,4 +1,0 @@
-package com.lucy.arti.winner.repository;
-
-public class WinnerRepositoryImpl implements WinnerRepositoryCustom{
-}
